refactor(web): move usePollingSuspension into polling zone
The use-polling-suspension.ts hook belongs architecturally with its
primary dependency (polling-state.ts). Keeping it in hooks/ caused
the test file (grouped into the polling zone by filename pattern) to
import across a zone boundary (polling → web), producing catastrophic
coupling metrics (0.75) and low cohesion (0.25) for the polling zone.

Changes:
- Add packages/web/src/viewer/polling/use-polling-suspension.ts (real impl)
- Export from polling/index.ts
- Replace hooks/use-polling-suspension.ts with a thin re-export shim
  (preserves backward compatibility for hooks barrel consumers)
- Update test to import directly from the polling zone
